Extract surface and wall geometry module
Move the helpers that turn 2D footprints into oriented 3D surface triangles,
clip existing surfaces to a new footprint, and rebuild the walls that close a
clipped cell into surface_geometry, along with the cardinal wall builders and
the cell bottom quad constructor.

These know about surfaces and walls but not about cells, nudge plans, or
rasters, so ordinary cell creation and the nudge repair passes can share them
without either importing the other.

Emitted meshes are unchanged: the four smoke configs stay byte identical.
